refactor: extract LSP, conform, blink, treesitter into plugin files; rewrite Section 10 with clean requires
